# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.GiveRewardRequest:
    properties:
      idea_id:
        type: string
      reward_type_id:
        type: string
    required:
    - idea_id
    - reward_type_id
    type: object
...
  dto.RewardResponse:
    properties:
      coffee_shop_id:
        type: string
      created_at:
        type: string
      given_at:
        type: string
      id:
        type: string
      idea_id:
        type: string
      is_activated:
        type: boolean
      receiver_id:
        type: string
      reward_type_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /admin/rewards:
    post:
      consumes:
      - application/json
      description: Creates a reward for the author of a specific idea. The caller
        must be the owner of the coffee shop.
      parameters:
      - description: Information about the reward to be given
        in: body
        name: reward_info
        required: true
        schema:
          $ref: '#/definitions/dto.GiveRewardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RewardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Give a reward
      tags:
      - rewards
  /admin/rewards/{id}:
    delete:
      description: Deletes a reward by its ID. The caller must be the owner of the
        coffee shop.
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Revoke a reward
      tags:
      - rewards
...
  /coffee-shops/{id}/rewards:
    get:
      description: Retrieves a paginated list of rewards associated with a specific
        coffee shop. The caller must be the owner.
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RewardResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get rewards for a coffee shop
      tags:
      - rewards
...
  /ideas/{id}/like:
    post:
      description: Like an idea by its ID
      parameters:
      - description: Idea ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like an idea
      tags:
      - likes
  /ideas/{id}/unlike:
    delete:
      description: Unlike an idea by its ID
      parameters:
      - description: Idea ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unlike an idea
      tags:
      - likes
...
  /rewards/{id}:
    get:
      description: Retrieves details of a single reward by its ID.
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RewardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a reward
      tags:
      - rewards
...
  /users/me/rewards:
    get:
      description: Retrieves a paginated list of rewards the currently authenticated
        user has received.
      parameters:
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RewardResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get my rewards
      tags:
      - rewards
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"